Fixed my service classes to have better logic. Only the handler touches the res.status, and the authservice has a purpose now. all tests passing

+++ server/src/main/java/service/UserService.java
@@ -4,7 +4,6 @@ import java.util.UUID;
 import dataaccess.*; // * imports all from that package
 import model.*;
 import requests.LoginRequest;
-import requests.LogoutRequest;
 import requests.RegisterRequest;
 import results.LoginResult;
 import results.SimpleResult;
@@ -13,10 +12,12 @@ public class UserService {
 
     private final UserDAO userDao;
     private final AuthDAO authDao;
+    private final AuthService authService;
 
-    public UserService(UserDAO userDao, AuthDAO authDao) {
+    public UserService(UserDAO userDao, AuthDAO authDao, AuthService authService) {
         this.userDao = userDao;
         this.authDao = authDao;
+        this.authService = authService;
     }
 
     public LoginResult register(RegisterRequest registerRequest) {
@@ -51,24 +52,32 @@ public class UserService {
         return new LoginResult(null, loginRequest.username(), authdata.authToken());
     }
 
-    public SimpleResult logout(LogoutRequest lreq) {
-        AuthData authData = authDao.findAuthData(lreq.authToken());
-        if (authData == null) {
-            return new SimpleResult("Error: unauthorized");
+    public SimpleResult logout(String authToken) {
+        String verification = authService.verifyAuth(authToken,false, null);
+        // bad path
+        if (verification.contains("Error")) {
+            return new SimpleResult(verification);
         }
-        authDao.deleteAuthData(authData);
-        return new SimpleResult(null);
+        // good path
+        if (verification.contains("verified")) {
+            // now we know that the token is legit. let's proceed
+            AuthData authData = authDao.findAuthData(authToken);
+            if (authData == null) {
+                return new SimpleResult("Error: unauthorized");
+            }
+            authDao.deleteAuthData(authData);
+            return new SimpleResult(null);
+        }
+        // wacky path
+        else {
+            return new SimpleResult("Error: an unexpected error occurred");
+        }
+
+
     }
 
     public void userClear() {
         userDao.clear();
     }
 
-    public void authClear() {
-        authDao.clear();
-    }
-
-    public AuthData getAuthData(String authToken) {
-        return authDao.findAuthData(authToken);
-    }
 }
